Fixed compiler warnings (vs2010)
git-svn-id: http://svn.ta3d.org/yuni/trunk@2126 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/stream.hxx
@@ -212,14 +212,14 @@ namespace File
 	template<class U>
 	inline uint Stream::write(const U& buffer)
 	{
-		return StreamTraitsWrite<Traits::CString<U>::valid, U>::Perform(pFd, buffer);
+		return (uint) StreamTraitsWrite<Traits::CString<U>::valid, U>::Perform(pFd, buffer);
 	}
 
 
 	template<class U>
 	inline uint Stream::write(const U& buffer, uint size)
 	{
-		return StreamTraitsWrite<Traits::CString<U>::valid, U>::Perform(pFd, buffer, size);
+		return (uint) StreamTraitsWrite<Traits::CString<U>::valid, U>::Perform(pFd, buffer, size);
 	}
